Fix: use Menu 'menu' prop, not 'name'; support w-sp in menu props
- 'name' is a separate 'identifier'-like Menu property. So use ":menu"
to get the menu entry 'name' (not to be confused with the menu 'name'
property!)
- Allow white space in menu entry 'name' (Example:
https://discourse.gohugo.io/t/working-with-multiple-menus-and-iterating-over-the-whole-set/7503)
- Refactor: menu-name(-str) -> menu-entry(-str) to prevent confusing
that with the menu name property.
- Remove 'children' from the list of settable Menu properties. As its
datatype is "Menu" and not "string" (from docs), I don't see that as
something a user would set. I believe, a user setting a menu parent
will mirror that menu entry as a child automatically in that parent
entry.
- Allow inheritance of HUGO_FRONT_MATTER_FORMAT property too.
- Add examples of white space use in ":menu" and ":name" for TOML and
YAML front matter