Wrap org-back-to-heading in ignore-errors
The rest of the change in this commit is indentation change and a very
old .md file that should have been deleted once the noexport feature
was implemented.

Use of ignoring errors thrown by org-back-to-heading:

You can add this to the posts Org file:

    # Local Variables:
    # eval: (add-hook 'after-save-hook #'org-hugo-publish-subtree :append :local)
    # End:

.. and have the post subtree auto-export on each save, but also not
throw an error if you happen to save the file while the point was
above the top-most Org headline in the file.