Use org-map-entries for index computation of a post
The index computation of a post can be really slow in a big blog for a
subtree at the bottom of the tree.  This commit implements the index
computation using `org-map-entries` making it much faster for the
average case for a large-ish blog, but slightly slower for the initial
few posts in the blog.