Remove HUGO_STATIC_IMAGE; Make attachment rewrite trigger correctly
Earlier, the images/ directory was given too much prominence... all
users might not have just an "images/" dir in the "static/" dir. A
user could have multiple dirs inside their static/ dir.. but earlier
the rewrite only used to happen to static/<HUGO_STATIC_IMAGE>. That
did not look right.

This commit fixes that.

Now ox-hugo tries to be smarter in figuring out where to copy the
outside-Hugo-static-dir files to.