Require text to be wrapped in braces to export as sub/superscript
Now ox-hugo by default requires text, to be sub/super-scripted, to
be wrapped in {}.

So now "a_b" will be exported as "a_b", but "a_{b}" will be exported
as "a<sub>b</sub>". To revert back to the earlier behavior, user needs
to add "#+OPTIONS: ^:t" to their Org file.