Remove Hugo version from default value of creator string
It doesn't make sense to encode the Hugo version at the time of the
Markdown file generation, because the generation of that has almost no
relation to the then Hugo version.

And the Hugo version is anyways encoded in the generator meta tags in
the HTML pages (unless a theme intentionally chooses to not to do so).