Fix quite a few Org keywords that couldn't be merged
For many keywords like #+HUGO_TAGS, it is useful to merge multiple
values when this is done:

    #+HUGO_TAGS: value1
    #+HUGO_TAGS: value2

This merging did not happen as expected for these keywords; this
commit fixes that.

    #+AUTHOR:
    #+HUGO_CUSTOM_FRONT_MATTER:
    #+HUGO_BLACKFRIDAY:
    #+HUGO_MENU:
    #+HUGO_MENU_OVERRIDE:
    #+HUGO_OUTPUTS:
    #+HUGO_RESOURCES:

Add test case for the same.

(Note that this fix was needed only for Org keyword values merging, not
subtree property merging.)