Use link attribute in figure when suitable
The Markdown form [{{< figure .. >}}](some link) results in wrapping
the <figure> in <a> in the final HTML, which is invalid.

- [X] Non-inlined hyperlinked images -- Use link in figure

https://github.com/kaushalmodi/ox-hugo/issues/125

One minor behavior change: Now <a> tag attributes like "target" and
"rel" cannot be applied *only to* the links that images hyperlink
to. The reason is to make the exported Markdown generate valid HTML,
and with how the "figure" shortcode is designed in HTML, those
parameters are not passed onto the <a> element in the generated
<figure> element.