Support passing target and rel attr to inline hyperlinked images
Also, now the target and rel attr are once again passed on to the
figure shortcode (a better solution to straightaway discarding those
as done in
https://github.com/kaushalmodi/ox-hugo/commit/e92fcf00375e9b20b4e9029b03fdc16409bd6f30).

Those attributes will still be ineffective as the internal figure
shortcode does not recognize those. But this can be fixed if Hugo adds
the support for those, or if a user simply overrides the internal
shortcode with their custom figure shortcode with that support added.