Add a newline after the id attr above named standalone images
This fixes the earlier generated invalid HTML.

---

Details:

Earlier, a figure with #+name:

    #+name: fig__unicorn
    [[/images/org-mode-unicorn-logo.png][file:/images/org-mode-unicorn-logo.png]]

would export as:

    <a id="org55ede7e"></a>
    {{< figure src="/images/org-mode-unicorn-logo.png" link="/images/org-mode-unicorn-logo.png" >}}

In HTML, that would look like:

    <p><a id="org55ede7e"></a>

        <figure>
            <a href="http://localhost:5555/images/org-mode-unicorn-logo.png">
                <img src="http://localhost:5555/images/org-mode-unicorn-logo.png" />
            </a>

        </figure>
    </p>

This resulted in invalid HTML as per the HTML5 validator.. because
figure is a block level element, and it cannot be put inside the p
tag.

The fix was to insert a blank line after the a tag so that it would be
in its own <p> element, and the following figure element would live
outside that.

With the fix in this commit,

    #+name: fig__unicorn
    [[/images/org-mode-unicorn-logo.png][file:/images/org-mode-unicorn-logo.png]]

now exports as:

    <a id="org55ede7e"></a>

    {{< figure src="/images/org-mode-unicorn-logo.png" link="/images/org-mode-unicorn-logo.png" >}}

In HTML, that now looks like:

    <p><a id="org55ede7e"></a></p>

    <figure>
        <a href="http://localhost:5555/images/org-mode-unicorn-logo.png">
            <img src="http://localhost:5555/images/org-mode-unicorn-logo.png" />
        </a>

    </figure>

.. which now is valid HTML.