Improve the logic for 1-level nested map custom front-matter
https://github.com/kaushalmodi/ox-hugo/issues/139

Remove the requirement to suffix such keys with "{}". This is an
improvement over the previous commit
https://github.com/kaushalmodi/ox-hugo/commit/d0a60c51049876fa15d9feed797907efcfa4829e.

---

The syntax for specifying nested map values custom front-matter is:

Using property drawer (in subtrees):

    :EXPORT_HUGO_CUSTOM_FRONT_MATTER: :PARENT_KEY '((KEY1 . VAL1) (KEY2 . VAL2) (KEY3 . VAL3) ..)

Using keyword:

    #+hugo_custom_front_matter: :PARENT_KEY '((KEY1 . VAL1) (KEY2 . VAL2) (KEY3 . VAL3) ..)

Note that the elements in the list *have to be* dotted-pairs
like (KEY1 . VAL1). That is what tells ox-hugo that this is a nested
map.

Only 1 level of nested keys is supported.