Add new defcustom org-hugo-tag-processing-functions
This commit reorganizes the former internal function
org-hugo--transform-org-tags into smaller functions and a new
defcustom org-hugo-tag-processing-functions. See its doc-string for
details.

This change is only internal, and the tag processing logic is exact
same as before (If this commit broke something for you, file a bug
report!).

This defcustom now allows users to add their own processing to the
tags e.g. prevent certain tag names (not the posts tagged with those!)
from getting exported, as requested in
https://github.com/kaushalmodi/ox-hugo/issues/142.

This commit now allows doing that using a small snippet like below:

    (defun org-hugo--tag-processing-fn-remove-tags-maybe (tags-list info)
      "Remove user-specified tags/categories.
    See `org-hugo-tag-processing-functions' for more info."
      ;; Use tag/category string (including @ prefix) exactly as used in Org file.
      (let ((tags-categories-to-be-removed '("DONE"))) ;"my_tag" "@my_cat"
        (cl-remove-if (lambda (tag_or_cat)
                        (member tag_or_cat tags-categories-to-be-removed))
                      tags-list)))
    (add-to-list 'org-hugo-tag-processing-functions
                 #'org-hugo--tag-processing-fn-remove-tags-maybe)