Make the "Table of Contents" heading a div element instead of h2
Hard-coding the "Table of Contents" heading to h2 now looked like a
bad idea.. it seems semantically incorrect. So now it is a div tag.

This also allows more flexibility on how to show that "Table of
Contents" heading.  See
https://ox-hugo.scripter.co/doc/org-toc/#table-of-contents-heading
for more info.

This is a "breaking change" in the sense that that heading will no
longer auto-render looking like a heading. So user might need to add
some CSS to make it look just like before. See example CSS in the
above doc link.