Changed variable assignment operator from = to :=
Running make doc md resulted in a lot of errors about unknown
variables from the examples layout. I changed the assignment operator
of those variables to := so that they would be declared before being
assigned a value. With this change all of the tests are now passing
for me, but I am not sure if this is what was intended or not.