Add <!--list-separator--> before headings that converted to li items
This is to cover the corner case where the previous section could be
ending in a plain list, and we want the current
heading-turned-list-item to merge that that list.

This is possible is the current heading level is 6 or more, or if the
`h' Export Option is set to a lower value by the user.

Related: https://github.com/kaushalmodi/ox-hugo/issues/344