fix: update org-id-locations if it's nil or empty
Note: The `org-id-locations` stores in ~/.emacs.d/.org-id-locations,
but there is no ~/.emacs.d exists when GitHub Workflow is running,
and that cause read & update org-id-locations failed.
So add an empty .emacs.d directory in test/ and then point the HOME to
test/ in Makefile.

Also use `hash-table-count' to instead `hash-table-empty-p' because
`hash-table-empty-p' is not available in Emacs 24.5.