Src blocks use code fence instead of `highlight` if Goldmark enabled
Ref:
- https://github.com/kaushalmodi/ox-hugo/pull/499
- https://github.com/kaushalmodi/ox-hugo/issues/305

This update also "fixes"
https://github.com/kaushalmodi/ox-hugo/issues/161.

The code fenced block don't have the same issue as the `highlight`
shortcode does. As we now export line numbered code blocks as fenced
code blocks by default and are moving away from the buggy `highlight`
shortcode, we are bypassing issue
https://github.com/kaushalmodi/ox-hugo/issues/161.

If you don't like the change in behavior of src block exports in this
commit, customize `org-hugo-goldmark`. See
https://github.com/kaushalmodi/ox-hugo/discussions/485.