Use <span> tags instead of <a> for target/radio link destinations
It makes more sense to use <span> tags instead of <a> if those tags
are to serve at link targets and not the actual hyperlinks.

Reason: Most of the CSS will blue underline an HTML element like

    <a id="foo">bar</a>

which would be confusing because those are not actually links.

This commit will instead create

    <span id="foo">bar</span>

This commit also adds a class attribute to the <<target>> and <<<radio
target>>> type link targets (where these span tags are created).