Fix export when "':t" export option is used
From the tests, it's quickly evident that Hugo/Goldmark do a better
job at smart quote conversions.

Org mode's smart quote conversion is always converting single quotes
to &rsquo;. For example, 'abc' will convert to &rsquo;abc&rsquo;. If
that's the desired behavior, enable it by using

    :EXPORT_OPTIONS: ':t

Goldmark converts 'abc' to &lsquo;abc&rsquo;. This is the default
behavior because ox-hugo ignores the user-set value of
`org-export-with-smart-quotes' due to the above reason.