Switch test and doc sites to use Hugo Modules
Now users who want to clone this repo and try building the test and
doc sites locally will need to have Go installed on their machine.

This change makes the Hugo theme and component dependency management
very convenient as it takes away the need to have submodules in this
repo.