Fix: Don't use `org-element-at-point` for Org stable based tests
Note: As of today, the Org stable or bugfix version's (9.5.2)
`org-element-at-point' returns the Org element at point *but* without
any of the inherited properties!!

So `org-element-map' is used where property inheritance needs to be
tested (because that does do the prop inheritance as expected!). This
issue doesn't exist in the `main' branch version of
`org-element-at-point'.