fix: compat: for `org-element-at-point` diff in Org 9.5 and older
In Org 9.5 and older versions, `org-element-at-point` returns an Org
element *without* the properties inherited from its parent headings
e.g. EXPORT_HUGO_BUNDLE.

In Org 9.6+ (main branch as of 2022-03-17), the returned element has
the inherited properties.

In this commit a `org-hugo--search-prop-in-parents` function is added
that allows searching for the specified property in parent headings.