fix!: html validation: Do not export <style> elements above TOC
<style> elements are not allowed inside <article> HTML elements. It is
very likely that the TOC + post content would be wrapped in an
<article> element, and in those cases, the HTML Validator will throw
an error like:

> Element style not allowed as child of element article in this
> context.

The breaking change will be seen only when both toc and section
numbering is enabled. The 'breakage' is visual -- that the numbered
headings in the TOCs will have the unordered list bullets.

To fix it, add this to the CSS:

    .toc.has-section-numbers ul {
      list-style: none;
    }