perf: Better way to auto-inject bibliography heading
Earlier method wasn't elegant because it relied on an empirically
derived regexp. The reason for this change was that this regexp was
falsely matching a <style> elsewhere in one of my blog posts (Nim
notes), and it caused a regex out of stack error :|

By advising `org-cite-export-bibliography` directly, we will always be
able to perfectly inject the heading before the bibliography. As a
side-effect, this approach now is able to inject the heading for
non-CSS i.e. basic style bibliography as well.