fix: Remove the blank line inserted between code blocks and captions
This change is part of the fix for
https://github.com/kaushalmodi/ox-hugo/issues/645. The good thing is
that there's no functional change in how Hugo/GoldMark (tested on Hugo
v0.99.1) render these Markdown files after the change.

So, to the user, it will be just a cosmetic change, that now there are
no blank lines between code blocks and their captions.