feat: Toggle ommission of subheading content
Previously, when a heading was rendered, each of its subheadings'
output was included in the output. With this change, a customizable
variable is added that, if set, will omit those subheadings that have
the :EXPORT_FILE_NAME property set from the output of their parent.

This can be handy for non-blogging sites, where the content is still
hierarchical, but doesn't lend itself to the blogging format where a
parent page includes the content of each of its children (like a year
index page, month index page, etc.)

At the rist of being annoyingly verbose, before this change, if you
had a post structure like this (in pseudo-org-mode format):

    * Post A
      :EXPORT_FILE_NAME: post-a

      Some text.

    ** Post B
       :EXPORT_FILE_NAME: post-b

       Text that relates to Post B.

    ** Post C

       General text, but no :EXPORT_FILE_NAME.

When rendered into HTML at the URL of /posts/post-a/, one would see:

    Post A

    Some text.

    Post B

    Text that relates to Post B.

    Post C

    General text, but no :EXPORT_FILE_NAME.

    …

After this change (and when the variable is set):

    Post A

    Some text.

    Post C

    General text, but no :EXPORT_FILE_NAME.

    …

The premise is that Post B will have its own page (as indicated by the
:EXPORT_FILE_NAME property) and so doesn't or shouldn't be included in
the content of Post A.